minor edits, typized services and controllers

+++ src/controllers/TasksControllers.ts
@@ -1,24 +1,30 @@
 import tasksServices from '../services/TasksServices';
+import {Request, Response} from 'express';
 import { validationResult } from "express-validator";
 import Sentry from "@sentry/node";
 
-require('dotenv').config();
+interface IRequestTasks extends Request {
+    userId?: string
+}
 
 class TasksControllers {
 
-    async getTasks(req: any, res: any) {
+    async getTasks(req: IRequestTasks, res: Response) {
         try {
             const errors = validationResult(req);
             if (!errors.isEmpty()) {
                 return res.status(400).json({ errors: errors.array() });
+            }
+            if(!req.userId){
+                return res.status(400).json({ errors: 'token does not contain user ID' });
             }
             await tasksServices.getTasks(req.userId)
                 .then((tasks: any) => {
                     res.status(200).send(tasks);
                 })
                 .catch(() => {
-                    res.status(404).send({
-                        errors: 'not found'
+                    res.status(400).send({
+                        errors: 'bad request'
                     });
                 })
         } catch (error) {
@@ -27,11 +33,14 @@ class TasksControllers {
 
     }
 
-    async createTask(req: any, res: any) {
+    async createTask(req: IRequestTasks, res: Response) {
         try {
             const errors = validationResult(req);
             if (!errors.isEmpty()) {
                 return res.status(400).json({ errors: errors.array() });
+            }
+            if(!req.userId){
+                return res.status(400).json({ errors: 'token does not contain user ID' });
             }
             const { title, isCompleted } = req.body;
             await tasksServices.createTask(title, isCompleted, req.userId)
@@ -49,11 +58,14 @@ class TasksControllers {
         }
     }
 
-    async updateTitle(req: any, res: any) {
+    async updateTitle(req: IRequestTasks, res: Response) {
         try {
             const errors = validationResult(req);
             if (!errors.isEmpty()) {
                 return res.status(400).json({ errors: errors.array() });
+            }
+            if(!req.userId){
+                return res.status(400).json({ errors: 'token does not contain user ID' });
             }
             const { title } = req.body;
             return await tasksServices.updateTitle(title, req.userId, req.params.id)
@@ -70,11 +82,14 @@ class TasksControllers {
         }
     }
 
-    async updateStatus(req: any, res: any) {
+    async updateStatus(req: IRequestTasks, res: Response) {
         try {
             const errors = validationResult(req);
             if (!errors.isEmpty()) {
                 return res.status(400).json({ errors: errors.array() });
+            }
+            if(!req.userId){
+                return res.status(400).json({ errors: 'token does not contain user ID' });
             }
             await tasksServices.updateStatus(req.userId, req.params.id)
                 .then((updatedTask) => {
@@ -88,11 +103,14 @@ class TasksControllers {
         }
     }
 
-    async deleteTask(req: any, res: any) {
+    async deleteTask(req: IRequestTasks, res: Response) {
         try {
             const errors = validationResult(req);
             if (!errors.isEmpty()) {
                 return res.status(400).json({ errors: errors.array() });
+            }
+            if(!req.userId){
+                return res.status(400).json({ errors: 'token does not contain user ID' });
             }
             res.send(await tasksServices.deleteTask(req.userId, req.params.id));
         } catch (error) {
